+++ CHANGES.md
@@ -16,7 +16,10 @@ Unreleased
 - h2: Track SETTINGS frames that haven't been acknowledged by the peer
   ([#22](https://github.com/anmonteiro/ocaml-h2/pull/22))
 - h2: Don't treat `CONNECT` requests as malformed
-  ([#32](https://github.com/anmonteiro/ocaml-h2/pull/32))
+  ([#32](https://github.com/anmonteiro/ocaml-h2/pull/32),
+  [#34](https://github.com/anmonteiro/ocaml-h2/pull/34))
+- h2: Fix frame writer not respecting the MAX\_FRAME\_SIZE setting
+  ([#34](https://github.com/anmonteiro/ocaml-h2/pull/34))
 
 0.1.0 2019-03-27
 --------------